+++ aim/web/api/runs/pydantic_models.py
@@ -13,14 +13,11 @@ class EncodedNumpyArray(BaseModel):
 
 class TraceBase(BaseModel):
     context: dict
-    metric_name: str
+    name: str
 
 
 class TraceOverview(TraceBase):
     last_value: float = 0.1
-
-
-RunTracesApiOut = List[TraceOverview]
 
 
 class TraceBaseView(TraceBase):
@@ -73,7 +70,7 @@ class MetricSearchRunView(BaseModel):
 
 class RunInfoOut(BaseModel):
     params: dict
-    traces: List[TraceOverview]
+    traces: Dict[str, List[TraceOverview]]
     props: PropsView
 
 
@@ -92,7 +89,7 @@ RunSearchApiOut = Dict[str, RunSearchRunView]
 # request models
 class AlignedTraceIn(BaseModel):
     context: dict
-    metric_name: str
+    name: str
     slice: Tuple[int, int, int]
 
 
@@ -145,3 +142,39 @@ class QuerySyntaxErrorOut(BaseModel):
         line: int
         offset: int
     detail: SE
+
+
+# image search API response models
+class RangeInfo(BaseModel):
+    records_range: Tuple[int, int]
+    index_range: Tuple[int, int]
+    record_slice: Tuple[int, int, int]
+    index_slice: Tuple[int, int, int]
+
+
+class ImageInfo(BaseModel):
+    caption: str
+    width: int
+    height: int
+    blob_uri: bytes
+    index: int
+
+
+class ImageSequenceFullView(TraceBase):
+    values: List[List[ImageInfo]]
+    iters: List[int]
+    epochs: List[int]
+    timestamps: List[float]
+
+
+class ImagesSearchRunView(BaseModel):
+    params: dict
+    traces: List[ImageSequenceFullView]
+    ranges: RangeInfo
+    props: PropsView
+
+
+RunImagesSearchApiOut = Dict[str, ImagesSearchRunView]
+
+
+URIBatchIn = List[str]