+++ aim/web/ui/src/components/BookmarkForm/BookmarkForm.tsx
@@ -12,7 +12,7 @@ import {
 import {
   IBookmarkFormState,
   IBookmarkFormProps,
-} from 'types/pages/metrics/components/BookmarkForm/BookmarkForm';
+} from 'types/components/BookmarkForm/BookmarkForm';
 
 const initialState: IBookmarkFormState = { name: '', description: '' };
 function BookmarkForm({