+++ aim/web/ui/src/services/models/explorer/createAppModel.ts
@@ -62,6 +62,7 @@ import {
   IRun,
   IMetricTrace,
   IParamTrace,
+  ISequence,
 } from 'types/services/models/metrics/runModel';
 import { IModel } from 'types/services/models/model';
 import {
@@ -133,7 +134,6 @@ import setAggregationEnabled from 'utils/app/setAggregationEnabled';
 import toggleSelectAdvancedMode from 'utils/app/toggleSelectAdvancedMode';
 import updateColumnsWidths from 'utils/app/updateColumnsWidths';
 import updateSortFields from 'utils/app/updateTableSortFields';
-import updateUrlParam from 'utils/app/updateUrlParam';
 import contextToString from 'utils/contextToString';
 import {
   ChartTypeEnum,
@@ -155,7 +155,6 @@ import getClosestValue from 'utils/getClosestValue';
 import getObjectPaths from 'utils/getObjectPaths';
 import getSmoothenedData from 'utils/getSmoothenedData';
 import getStateFromUrl from 'utils/getStateFromUrl';
-import getUrlWithParam from 'utils/getUrlWithParam';
 import JsonToCSV from 'utils/JsonToCSV';
 import { SmoothingAlgorithmEnum } from 'utils/smoothingData';
 import { getItem, setItem } from 'utils/storage';
@@ -439,7 +438,7 @@ function createAppModel({
       }
     }
 
-    function updateData(newData: IRun<IMetricTrace>[]): void {
+    function updateData(newData: ISequence<IMetricTrace>[]): void {
       const configData = model.getState()?.config;
       if (configData) {
         setModelData(newData, configData);
@@ -476,11 +475,19 @@ function createAppModel({
             }
             if (components.table) {
               state.tableData = [];
+              state.config = {
+                ...configData,
+                table: {
+                  ...configData?.table,
+                  resizeMode: ResizeModeEnum.Resizable,
+                },
+              };
             }
 
             model.setState({
               requestIsPending: false,
               queryIsEmpty: true,
+
               ...state,
             });
           } else {
@@ -632,7 +639,7 @@ function createAppModel({
               run: moment(metric.run.props.creation_time * 1000).format(
                 'HH:mm:ss · D MMM, YY',
               ),
-              metric: metric.metric_name,
+              metric: metric.name,
               context: contextToString(metric.context)?.split(',') || [''],
               value:
                 closestIndex === null
@@ -804,7 +811,7 @@ function createAppModel({
       model.setState({ config: configData });
     }
 
-    function processData(data: IRun<IMetricTrace>[]): {
+    function processData(data: ISequence<IMetricTrace>[]): {
       data: IMetricsCollection<IMetric>[];
       params: string[];
       highLevelParams: string[];
@@ -818,7 +825,7 @@ function createAppModel({
       let contexts: string[] = [];
       const paletteIndex: number = configData?.grouping?.paletteIndex || 0;
 
-      data?.forEach((run: IRun<IMetricTrace>) => {
+      data?.forEach((run: ISequence<IMetricTrace>) => {
         params = params.concat(getObjectPaths(run.params, run.params));
         highLevelParams = highLevelParams.concat(
           getObjectPaths(run.params, run.params, '', false, true),
@@ -851,7 +858,7 @@ function createAppModel({
             }
             const metricKey = encode({
               runHash: run.hash,
-              metricName: trace.metric_name,
+              metricName: trace.name,
               traceContext: trace.context,
             });
             return createMetricModel({
@@ -905,7 +912,7 @@ function createAppModel({
       shouldURLUpdate?: boolean,
     ): void {
       const { data, params, highLevelParams, contexts } = processData(
-        model.getState()?.rawData as IRun<IMetricTrace>[],
+        model.getState()?.rawData as ISequence<IMetricTrace>[],
       );
       const groupingSelectOptions = [
         ...getGroupingSelectOptions({
@@ -974,7 +981,7 @@ function createAppModel({
     }
 
     function setModelData(
-      rawData: IRun<IMetricTrace>[],
+      rawData: ISequence<IMetricTrace>[],
       configData: IMetricAppConfig,
     ): void {
       const sortFields = model.getState()?.config?.table?.sortFields;
@@ -1493,7 +1500,7 @@ function createAppModel({
         type: 'text/csv;charset=utf-8;',
       });
       saveAs(blob, `${appName}-${moment().format('HH:mm:ss · D MMM, YY')}.csv`);
-      analytics.trackEvent(`[${appName}] Export runs data to CSV`);
+      analytics.trackEvent(`[${appName}Explore] Export runs data to CSV`);
     }
 
     const onActivePointChange = _.debounce(
@@ -1655,6 +1662,11 @@ function createAppModel({
         },
       });
       setItem('metricsLUConfig', encode(newLiveUpdateConfig));
+      analytics.trackEvent(
+        `[${appName}Explorer] Switch live-update ${
+          config.enabled ? 'on' : 'off'
+        }`,
+      );
     }
 
     function destroy(): void {
@@ -1745,13 +1757,13 @@ function createAppModel({
         onMetricsSelectChange<D>(
           data: D & Partial<ISelectMetricsOption[]>,
         ): void {
-          onMetricsSelectChange({ data, model, appName });
+          onMetricsSelectChange({ data, model });
         },
         onSelectRunQueryChange(query: string): void {
-          onSelectRunQueryChange({ query, model, appName });
+          onSelectRunQueryChange({ query, model });
         },
         onSelectAdvancedQueryChange(query: string): void {
-          onSelectAdvancedQueryChange({ query, model, appName });
+          onSelectAdvancedQueryChange({ query, model });
         },
         toggleSelectAdvancedMode(): void {
           toggleSelectAdvancedMode({ model, appName });
@@ -1919,21 +1931,8 @@ function createAppModel({
         // }
         model.setState({ ...state });
         if (!appId) {
-          const searchParam = new URLSearchParams(window.location.search);
-          const searchFromUrl = searchParam.get('search');
-          const urlFromStorage = getItem(`${appName}Url`);
-          if (searchFromUrl) {
-            setItem(
-              `${appName}Url`,
-              getUrlWithParam({ search: searchFromUrl }),
-            );
-          } else {
-            if (urlFromStorage) {
-              window.history.pushState(null, '', urlFromStorage);
-            }
-          }
+          setDefaultAppConfigData();
         }
-        setDefaultAppConfigData();
 
         const liveUpdateState = model.getState()?.config.liveUpdate;
 
@@ -1948,7 +1947,10 @@ function createAppModel({
         return getRunsData();
       }
 
-      function getRunsData(isInitial = true): {
+      function getRunsData(
+        shouldUrlUpdate?: boolean,
+        isInitial = true,
+      ): {
         call: () => Promise<void>;
         abort: () => void;
       } {
@@ -1969,6 +1971,10 @@ function createAppModel({
           pagination?.limit,
           pagination?.offset,
         );
+
+        if (shouldUrlUpdate) {
+          updateURL({ configData, appName });
+        }
 
         return {
           call: async () => {
@@ -2155,9 +2161,9 @@ function createAppModel({
 
         data?.forEach((run: IRun<IParamTrace>, index) => {
           params = params.concat(getObjectPaths(run.params, run.params));
-          run.traces.forEach((trace) => {
-            metricsColumns[trace.metric_name] = {
-              ...metricsColumns[trace.metric_name],
+          run.traces.metric.forEach((trace) => {
+            metricsColumns[trace.name] = {
+              ...metricsColumns[trace.name],
               [contextToString(trace.context) as string]: '-',
             };
           });
@@ -2382,12 +2388,12 @@ function createAppModel({
           }
           metricsCollection.data.forEach((metric: any) => {
             const metricsRowValues = { ...initialMetricsRowData };
-            metric.run.traces.forEach((trace: any) => {
+            metric.run.traces.metric.forEach((trace: any) => {
               metricsRowValues[
                 `${
-                  isSystemMetric(trace.metric_name)
-                    ? trace.metric_name
-                    : `${trace.metric_name}_${contextToString(trace.context)}`
+                  isSystemMetric(trace.name)
+                    ? trace.name
+                    : `${trace.name}_${contextToString(trace.context)}`
                 }`
               ] = formatValue(trace.last_value.last);
             });
@@ -2398,8 +2404,10 @@ function createAppModel({
               color: metricsCollection.color ?? metric.color,
               dasharray: metricsCollection.dasharray ?? metric.dasharray,
               experiment: metric.run.props.experiment.name ?? 'default',
-              run: metric.run.props.name ?? '-',
-              metric: metric.metric_name,
+              run: moment(metric.run.props.creation_time * 1000).format(
+                'HH:mm:ss · D MMM, YY',
+              ),
+              metric: metric.name,
               ...metricsRowValues,
             };
             rowIndex++;
@@ -2486,7 +2494,7 @@ function createAppModel({
               },
             },
           });
-          return getRunsData(false);
+          return getRunsData(false, false);
         }
       }
 
@@ -2552,7 +2560,9 @@ function createAppModel({
           type: 'text/csv;charset=utf-8;',
         });
         saveAs(blob, `runs-${moment().format('HH:mm:ss · D MMM, YY')}.csv`);
-        analytics.trackEvent('[RunsExplorer][Table] Export runs data to CSV');
+        analytics.trackEvent(
+          `[${appName}Explore][Table] Export runs data to CSV`,
+        );
       }
 
       function onModelNotificationDelete(id: number): void {
@@ -2637,6 +2647,11 @@ function createAppModel({
         });
 
         setItem('runsLUConfig', encode(newLiveUpdateConfig));
+        analytics.trackEvent(
+          `[${appName}Explorer] Switch live-update ${
+            config.enabled ? 'on' : 'off'
+          }`,
+        );
       }
 
       const methods = {
@@ -2656,13 +2671,7 @@ function createAppModel({
       if (selectForm) {
         Object.assign(methods, {
           onSelectRunQueryChange(query: string): void {
-            onSelectRunQueryChange({ query, model, appName });
-          },
-          updateSelectStateUrl(): void {
-            const selectData = model.getState()?.config?.select;
-            if (selectData) {
-              updateUrlParam({ data: { search: selectData }, appName });
-            }
+            onSelectRunQueryChange({ query, model });
           },
         });
       }
@@ -2847,6 +2856,13 @@ function createAppModel({
               }
               if (components.table) {
                 state.tableData = [];
+                state.config = {
+                  ...configData,
+                  table: {
+                    ...configData?.table,
+                    resizeMode: ResizeModeEnum.Resizable,
+                  },
+                };
               }
 
               model.setState({
@@ -2962,12 +2978,12 @@ function createAppModel({
 
             metricsCollection.data.forEach((metric: any) => {
               const metricsRowValues = { ...initialMetricsRowData };
-              metric.run.traces.forEach((trace: any) => {
+              metric.run.traces.metric.forEach((trace: any) => {
                 metricsRowValues[
                   `${
-                    isSystemMetric(trace.metric_name)
-                      ? trace.metric_name
-                      : `${trace.metric_name}_${contextToString(trace.context)}`
+                    isSystemMetric(trace.name)
+                      ? trace.name
+                      : `${trace.name}_${contextToString(trace.context)}`
                   }`
                 ] = formatValue(trace.last_value.last);
               });
@@ -2985,7 +3001,7 @@ function createAppModel({
                 run: moment(metric.run.props.creation_time * 1000).format(
                   'HH:mm:ss · D MMM, YY',
                 ),
-                metric: metric.metric_name,
+                metric: metric.name,
                 ...metricsRowValues,
               };
               rowIndex++;
@@ -3122,12 +3138,12 @@ function createAppModel({
                       };
                     }
                     if (type === 'metrics') {
-                      run.run.traces.forEach((trace: IParamTrace) => {
+                      run.run.traces.metric.forEach((trace: IParamTrace) => {
                         const formattedContext = `${
                           value?.param_name
                         }-${contextToString(trace.context)}`;
                         if (
-                          trace.metric_name === value?.param_name &&
+                          trace.name === value?.param_name &&
                           _.isEqual(trace.context, value?.context)
                         ) {
                           values[formattedContext] = trace.last_value.last;
@@ -3445,9 +3461,9 @@ function createAppModel({
           highLevelParams = highLevelParams.concat(
             getObjectPaths(run.params, run.params, '', false, true),
           );
-          run.traces.forEach((trace) => {
-            metricsColumns[trace.metric_name] = {
-              ...metricsColumns[trace.metric_name],
+          run.traces.metric.forEach((trace) => {
+            metricsColumns[trace.name] = {
+              ...metricsColumns[trace.name],
               [contextToString(trace.context) as string]: '-',
             };
           });
@@ -3748,6 +3764,11 @@ function createAppModel({
         });
 
         setItem('paramsLUConfig', encode(newLiveUpdateConfig));
+        analytics.trackEvent(
+          `[${appName}Explorer] Switch live-update ${
+            config.enabled ? 'on' : 'off'
+          }`,
+        );
       }
 
       function destroy(): void {
@@ -3825,12 +3846,11 @@ function createAppModel({
                 select: { ...configData.select, params: data },
               };
 
-              updateURL({ configData: newConfig, appName });
               model.setState({ config: newConfig });
             }
           },
           onSelectRunQueryChange(query: string): void {
-            onSelectRunQueryChange({ query, model, appName });
+            onSelectRunQueryChange({ query, model });
           },
         });
       }