+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
-## Unreleased
+## 3.1.0 Nov 20 2021
 
+- Add ability to explore tracked images (VkoHov)
 - Improve rendering performance by virtualizing table columns (roubkar)
 - Add ability to apply grouping by higher level param key (roubkar)
 - Add ability to specify repository path during `aim init` via `--repo` argument (rubenaprikyan)
+
+## 3.0.7 Nov 17 2021
+
+- Fix for missing metrics when numpy.float64 values tracked (alberttorosyan)
+
 ## 3.0.6 Nov 9 2021
 
 - Fix for blocking container optimization for in progress runs (alberttorosyan)